store includes in a hashmap(dashmap)
this commit introduces the dashmap crate which stores the includes
a hashmap so that they are read only once and only when they are
needed. done for performance reasons so that the file reads are
minimal but I don't think there's any noticable difference since
file reading is already pretty fast.